+++ Projects/Playground/Playground.Droid/Fragments/ChildView.cs
@@ -8,17 +8,20 @@ using Android.Views;
 using MvvmCross.Droid.Support.V4;
 using MvvmCross.Platforms.Android.Binding.BindingContext;
 using MvvmCross.Platforms.Android.Presenters.Attributes;
-using MvvmCross.Presenters;
-using MvvmCross.Presenters.Attributes;
-using MvvmCross.ViewModels;
 using Playground.Core.ViewModels;
 
-namespace Playground.Droid.Views
+namespace Playground.Droid.Fragments
 {
-    [MvxFragmentPresentation(typeof(RootViewModel), Resource.Id.content_frame)]
+    [MvxFragmentPresentation(typeof(RootViewModel), Resource.Id.content_frame, true,
+                             Resource.Animation.abc_fade_in,
+                             Resource.Animation.abc_fade_out,
+                             Resource.Animation.abc_fade_in,
+                             Resource.Animation.abc_fade_out)]
     [MvxFragmentPresentation(typeof(SplitRootViewModel), Resource.Id.split_content_frame)]
-    [Register(nameof(OverrideAttributeView))]
-    public class OverrideAttributeView : MvxFragment<OverrideAttributeViewModel>, IMvxOverridePresentationAttribute
+    [MvxFragmentPresentation(typeof(TabsRootViewModel), Resource.Id.content_frame)]
+    [MvxFragmentPresentation(fragmentHostViewType: typeof(ModalNavView), fragmentContentId: Resource.Id.dialog_content_frame)]
+    [Register(nameof(ChildView))]
+    public class ChildView : MvxFragment<ChildViewModel>
     {
         public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
         {
@@ -29,16 +32,9 @@ namespace Playground.Droid.Views
             return view;
         }
 
-        public MvxBasePresentationAttribute PresentationAttribute(MvxViewModelRequest request)
+        public override void OnDestroy()
         {
-            return new MvxFragmentPresentationAttribute(
-                typeof(RootViewModel),
-                Resource.Id.content_frame,
-                false,
-                Resource.Animation.abc_fade_in,
-                Resource.Animation.abc_fade_out,
-                Resource.Animation.abc_fade_in,
-                Resource.Animation.abc_fade_out);
+            base.OnDestroy();
         }
     }
 }
