+++ Projects/Playground/Playground.Droid/Fragments/ModalNavView.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using Android.OS;
 using Android.Runtime;
 using Android.Views;
@@ -10,17 +11,26 @@ using MvvmCross.Platforms.Android.Binding.BindingContext;
 using MvvmCross.Platforms.Android.Presenters.Attributes;
 using Playground.Core.ViewModels;
 
-namespace Playground.Droid.Views
+namespace Playground.Droid.Fragments
 {
-    [MvxFragmentPresentation(typeof(SplitRootViewModel), Resource.Id.split_content_frame)]
-    [Register(nameof(SplitDetailNavView))]
-    public class SplitDetailNavView : MvxFragment<SplitDetailNavViewModel>
+    [MvxDialogFragmentPresentation]
+    [Register(nameof(ModalNavView))]
+    public class ModalNavView : MvxDialogFragment<ModalNavViewModel>
     {
+        public ModalNavView()
+        {
+        }
+
+        protected ModalNavView(IntPtr javaReference, JniHandleOwnership transfer)
+            : base(javaReference, transfer)
+        {
+        }
+
         public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
         {
-            base.OnCreateView(inflater, container, savedInstanceState);
+            var ignore = base.OnCreateView(inflater, container, savedInstanceState);
 
-            var view = this.BindingInflate(Resource.Layout.SplitDetailView, null);
+            var view = this.BindingInflate(Resource.Layout.ModalNavView, null);
 
             return view;
         }
