+++ Projects/Playground/Playground.Droid/Fragments/SecondChildView.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MS-PL license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using Android.OS;
 using Android.Runtime;
 using Android.Views;
@@ -11,26 +10,18 @@ using MvvmCross.Platforms.Android.Binding.BindingContext;
 using MvvmCross.Platforms.Android.Presenters.Attributes;
 using Playground.Core.ViewModels;
 
-namespace Playground.Droid.Views
+namespace Playground.Droid.Fragments
 {
-    [MvxDialogFragmentPresentation]
-    [Register(nameof(NestedModalView))]
-    public class NestedModalView : MvxDialogFragment<NestedModalViewModel>
+    [MvxFragmentPresentation(typeof(RootViewModel), Resource.Id.content_frame, true)]
+    [MvxFragmentPresentation(typeof(SplitRootViewModel), Resource.Id.split_content_frame)]
+    [Register(nameof(SecondChildView))]
+    public class SecondChildView : MvxFragment<SecondChildViewModel>
     {
-        public NestedModalView()
-        {
-        }
-
-        protected NestedModalView(IntPtr javaReference, JniHandleOwnership transfer)
-            : base(javaReference, transfer)
-        {
-        }
-
         public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
         {
-            var ignore = base.OnCreateView(inflater, container, savedInstanceState);
+            base.OnCreateView(inflater, container, savedInstanceState);
 
-            var view = this.BindingInflate(Resource.Layout.ChildView, null);
+            var view = this.BindingInflate(Resource.Layout.SecondChildView, null);
 
             return view;
         }
