+++ Projects/Playground/Playground.Droid/Fragments/SharedElementSecondChildView.cs
@@ -3,32 +3,35 @@
 // See the LICENSE file in the project root for more information.
 
 using Android.OS;
-using Android.Runtime;
+using Android.Transitions;
 using Android.Views;
 using MvvmCross.Droid.Support.V4;
 using MvvmCross.Platforms.Android.Binding.BindingContext;
 using MvvmCross.Platforms.Android.Presenters.Attributes;
+using MvvmCross.Platforms.Android.Views;
 using Playground.Core.ViewModels;
 
-namespace Playground.Droid.Views
+namespace Playground.Droid.Fragments
 {
-    [MvxTabLayoutPresentation(TabLayoutResourceId = Resource.Id.tabs, ViewPagerResourceId = Resource.Id.viewpager, Title = "Tab 2", ActivityHostViewModelType = typeof(TabsRootViewModel))]
-    [MvxTabLayoutPresentation(TabLayoutResourceId = Resource.Id.tabs, ViewPagerResourceId = Resource.Id.viewpager, Title = "Tab 2", ActivityHostViewModelType = typeof(TabsRootBViewModel))]
-    [Register(nameof(Tab2View))]
-    public class Tab2View : MvxFragment<Tab2ViewModel>
+    [MvxFragmentPresentation(typeof(SharedElementRootViewModel), Resource.Id.shared_content_frame, true)]
+    public class SharedElementSecondChildView : MvxFragment<SharedElementSecondChildViewModel>
     {
         public override void OnCreate(Bundle savedInstanceState)
         {
             base.OnCreate(savedInstanceState);
 
-            // Create your fragment here
+            if (Build.VERSION.SdkInt >= BuildVersionCodes.Lollipop)
+            {
+                SharedElementEnterTransition = TransitionInflater.From(Activity).InflateTransition(Android.Resource.Transition.Move);
+            }
         }
 
         public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
         {
             base.OnCreateView(inflater, container, savedInstanceState);
 
-            var view = this.BindingInflate(Resource.Layout.Tab2View, null);
+            var view = this.BindingInflate(Resource.Layout.SharedElementSecondChildView, null);
+            Arguments.SetSharedElementsById(view);
 
             return view;
         }
