+++ Projects/Playground/Playground.Droid/Fragments/SplitDetailNavView.cs
@@ -10,24 +10,17 @@ using MvvmCross.Platforms.Android.Binding.BindingContext;
 using MvvmCross.Platforms.Android.Presenters.Attributes;
 using Playground.Core.ViewModels;
 
-namespace Playground.Droid.Views
+namespace Playground.Droid.Fragments
 {
-    [MvxTabLayoutPresentation(TabLayoutResourceId = Resource.Id.tabs, ViewPagerResourceId = Resource.Id.viewpager, Title = "Tab 3")]
-    [Register(nameof(Tab3View))]
-    public class Tab3View : MvxFragment<Tab3ViewModel>
+    [MvxFragmentPresentation(typeof(SplitRootViewModel), Resource.Id.split_content_frame)]
+    [Register(nameof(SplitDetailNavView))]
+    public class SplitDetailNavView : MvxFragment<SplitDetailNavViewModel>
     {
-        public override void OnCreate(Bundle savedInstanceState)
-        {
-            base.OnCreate(savedInstanceState);
-
-            // Create your fragment here
-        }
-
         public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
         {
             base.OnCreateView(inflater, container, savedInstanceState);
 
-            var view = this.BindingInflate(Resource.Layout.Tab3View, null);
+            var view = this.BindingInflate(Resource.Layout.SplitDetailView, null);
 
             return view;
         }
